% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWikifamilytree.R
\name{buildTreeGrid}
\alias{buildTreeGrid}
\title{Parse Tree}
\usage{
buildTreeGrid(tree_lines)
}
\arguments{
\item{tree_lines}{A character vector containing the lines of the tree structure.}
}
\value{
A data frame containing the tree structure.
}
\description{
Parse Tree
}
\examples{
buildTreeGrid(c(" | | GMa |~|y|~| GPa | ", " | | MOM |y| DAD | | "))
}
\keyword{internal}
