% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseLines.R
\name{extractInfoFromLines}
\alias{extractInfoFromLines}
\title{Extract Information from Lines by Tag}
\usage{
extractInfoFromLines(lines, tag)
}
\arguments{
\item{lines}{A character vector of GEDCOM lines to search through.}

\item{tag}{A character string representing the GEDCOM tag to look for (e.g
"DATE", "PLAC", "CAUS").}
}
\value{
A character string with the extracted information from the first matching line, or `NA_character_` if no matching line is found.
}
\description{
Given a set of lines (e.g., direct children of an event) and a
GEDCOM tag, this function searches for the first line that contains the tag as a whole word and extracts the relevant information using the `extractInfo()` function. If no matching line is found, it returns `NA_character_`.
}
\keyword{internal}
